fixed a bug in the GPU improper calculation where "acos()" was called instead of "acosf()"
git-svn-id: file:///Users/joaander/devel/hoomd-git-convert/repos/branches/hoomd-0.8@1984 fa922fa7-2fde-0310-acd8-f43f465a7996

+++ src/cuda/HarmonicImproperForceGPU.cu
@@ -207,7 +207,7 @@ extern "C" __global__ void gpu_compute_harmonic_improper_forces_kernel(gpu_force
 		float s = sqrt(1.0f - c*c);
 		if (s < SMALL) s = SMALL;
 
-		float domega = acos(c) - chi;
+		float domega = acosf(c) - chi;
 		float a = K * domega;
 
 		// calculate the energy, 1/4th for each atom
